Added random tests to 3 card functions and updated makefile to include new files
Tests are for adventurer, outpost, and village. MakeFile usedby typing make all and then make runRandomTests. The input may appear to hang on the 3rd test. As explained in my paper this is due to the Random() Number Generator. It is prompting the console for additional seeds so enter a random number like 3 or 8 so the program can continue running